requirements.txt
torch
torchvision
numpy
pillow
matplotlib
tqdm
scikit-learn
